Added Contacts  Authorisation

+++ app/controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require('../models/contactModel');
 
 //@desc GET all contacts
 //@route GET /api/contacts
-//@access Public
+//@access Private
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({user_id: req.user.id});
   res.status(200).json(contacts);
 })
 
 //@desc Create new contacts
 //@route POST /api/contacts
-//@access Public
+//@access Private
 
 const createContacts = asyncHandler(async (req, res) => {
   console.log("The request body: ", req.body);
@@ -26,7 +26,8 @@ const createContacts = asyncHandler(async (req, res) => {
   const contact = await Contact.create({
     name,
     email,
-    phone
+    phone,
+    user_id: req.user.id
   })
 
   res.status(201).json(contact); 
@@ -35,7 +36,7 @@ const createContacts = asyncHandler(async (req, res) => {
 
 //@desc get a contact
 //@route GET /api/contacts/:id
-//@access Public
+//@access Private
 
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -48,13 +49,18 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc update a contact
 //@route UPDATE /api/contacts/:id
-//@access Public
+//@access Private
 
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) { 
     res.status(404);
     throw new Error('Contact not found!');
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error('User not authorized!');
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, {
@@ -66,7 +72,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc delete a contact
 //@route DELETE /api/contacts/:id
-//@access Public
+//@access Private
 
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -74,6 +80,12 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error('Contact not found!');
   }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error('User not authorized!');
+  }
+  
   await contact.deleteOne();
 
   res.status(200).json(contact);
